src/requirements.txt
setuptools
pandas
numpy
scikit-learn
bentoml
confluent-kafka
python-dotenv
requests
cassandra-driver
asyncio
websocket-client
evidently
dvc
ta==0.10.1
pmdarima
feast
scipy
optuna
vectorbt
matplotlib